+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/event/ClientEvents.kt
@@ -2,6 +2,7 @@ package dev.aaronhowser.mods.excessive_utilities.event
 
 import dev.aaronhowser.mods.aaron.misc.AaronExtensions.isItem
 import dev.aaronhowser.mods.excessive_utilities.ExcessiveUtilities
+import dev.aaronhowser.mods.excessive_utilities.block_entity.ChandelierBlockEntity
 import dev.aaronhowser.mods.excessive_utilities.block_entity.DrumBlockEntity
 import dev.aaronhowser.mods.excessive_utilities.block_entity.SoundMufflerBlockEntity
 import dev.aaronhowser.mods.excessive_utilities.client.render.AngelWingRenderer
@@ -226,6 +227,8 @@ object ClientEvents {
 		event.registerBlockEntityRenderer(ModBlockEntityTypes.ENDER_COLLECTOR.get(), ::EnderCollectorBER)
 		event.registerBlockEntityRenderer(ModBlockEntityTypes.QED.get(), ::QedBER)
 		event.registerBlockEntityRenderer(ModBlockEntityTypes.CREATIVE_HARVEST.get(), ::CreativeHarvestBER)
+		event.registerBlockEntityRenderer(ModBlockEntityTypes.MAGNUM_TORCH.get(), ::MagnumTorchBER)
+		event.registerBlockEntityRenderer(ModBlockEntityTypes.CHANDELIER.get(), ::ChandelierBER)
 	}
 
 	@SubscribeEvent