+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fixed the color of the Nether Star Generator's particles from white to "bistre"
 - Generators can no longer accept energy from outside sources
 - Generators now push energy into adjacent blocks
+- When holding a Chandelier or Magnum Torch, it will render a box showing the area of effect
 
 ### Fixed